Fix build
Add jar
