+++ docker/withdrawal.dockerfile
@@ -2,7 +2,7 @@ FROM openjdk:8-jre
 
 WORKDIR /opt/notary
 
-COPY build/libs/notary-1.0-SNAPSHOT-all.jar /opt/notary/notary.jar
+COPY eth-withdrawal/build/libs/eth-withdrawal-all.jar /opt/notary/notary.jar
 COPY iroha_bindings/linux/* /opt/notary/
 ENV LD_LIBRARY_PATH="/opt/notary:${LD_LIBRARY_PATH}"
