+++ eth-vacuum/build.gradle
@@ -4,6 +4,8 @@ dependencies {
     compile project(":eth")
 }
 
+apply plugin: 'com.github.johnrengelman.shadow'
+
 sonarqube {
     properties {
         property "sonar.projectKey", "notary:eth-vacuum"
